+++ R/gt.R
@@ -30,8 +30,7 @@
 #'   group labels for generation of stub row groups. If the input `data` table
 #'   has the `grouped_df` class (through use of the [dplyr::group_by()] function
 #'   or associated `group_by*()` functions) then any input here is ignored.
-#' @param caption An optional table caption to use for cross-referencing
-#'   in R Markdown documents and **bookdown** book projects.
+#' @inheritParams tab_caption
 #' @param rownames_to_stub An option to take rownames from the input `data`
 #'   table as row captions in the display table stub.
 #' @param auto_align Optionally have column data be aligned depending on the
@@ -187,15 +186,7 @@ gt <- function(
   # we'll use the header from `tab_header` as the table caption); this might
   # require some more thought still because a `caption` arg might also be
   # sensible in `tab_header`
-  if (!is.null(caption)) {
-
-    data <-
-      dt_options_set_value(
-        data = data,
-        option = "table_caption",
-        value = caption
-      )
-  }
+  data <- tab_caption(data = data, caption = caption)
 
   # Apply the `gt_tbl` class to the object while
   # also keeping the `data.frame` class